Allowing non-standard names for `tsconfig.json`
This simply replaces the previous strict check (`name ===
"tsconfig.json"`) with a more relaxed check that simply ensures that
the name begins with `tsconfig` and ends in `.json`. This is required
in the case of e.g. Angular applications that use names like
`tsconfig.app.json` and `tsconfig.spec.json`.