Add --include flag: further limits files from tsconfig.json.
In a large project, it is slow and not needed to extract schemata from
all TypeScript files in `tsconfig.json` (also it can be infeasible due
to limitations of what can be converted to JSON schema).  OTOH listing
the files as the command-line argument ignores all the parameters from
`tsconfig.json`.  `--include` gets us the advantages of both.

Tests for --include flag (and programFromConfig).

(PR ready: no generated files).