Resolve type aliases
In some situations getSymbolAtLocation() returns an alias which does
not provide JSDoc comments. In this case the type must be resolved
with getAliasedSymbol first to be able to read comments for description
and annotations.

I also had to rename two schema files in the unique-names unit tests
because the symbol IDs are now different. I guess that's because now the
real symbols are used instead of the alias symbols. I hope this isn't
a breaking change, I'm not even sure what the use case for this unique
name feature is anyway.