Fix AJV warning, "$ref: keywords ignored in schema"
The improved handling for recursive schemas that was added in #383 resulted in type-related keywords being added alongside the `$ref` keywords. Such keywords are ignored according to draft-07 of the JSON Schema, so AJV (and possibly other tools) warn about them.

For more information, see https://github.com/ajv-validator/ajv/issues/1121

To address this, I revised the recursive ref logic to only include "annotation" keywords, based on the list here:

http://json-schema.org/understanding-json-schema/reference/generic.html#annotations

(That page also includes `title`, but typescript-json-schema currently doesn't recognize that.)

Note that draft 2019-09 of JSON Schema changes `$ref` to act like a normal keyword, so this change will no longer be necessary in that version of the spec.

Fixes #391.