Fix maxItems vs. additionalItems
If I understand correctly, typescript-json-schema's use of additionalItems was incorrect; it had the result of allowing any number of additional items to be added to the end of the tuple, instead of restricting the tuple to only the listed items. It was also causing warnings in AJV 7's new strict mode:

> strict mode: "items" is 2-tuple, but minItems or maxItems/additionalItems are not specified or different

For more information:

- https://github.com/ajv-validator/ajv/blob/v7.0.0/docs/strict-mode.md#prohibit-unconstrained-tuples
- https://json-schema.org/understanding-json-schema/reference/array.html#tuple-validation

Fixes #393.