Single union case shouldn't overwrite existing properties (#550)
In the case of a single union, we were just copying all properties
across. This can overwrite a pre-existing field, in particular a
more-specific description that has been specified on some including
type.

This came up in a project where enabling `strictNullChecks` caused a
bunch of descriptions to change for seemingly no reason.